requirements.txt
alembic==1.3.2
Authlib==0.13
autopep8==1.4.4
Babel==2.7.0
cachetools==2.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
ecdsa==0.14.1
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
get==2019.4.13
google-api-core==1.14.3
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.23.0
google-cloud-core==1.1.0
google-cloud-speech==1.3.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.6
oauth2client==4.1.3
openpyxl==3.0.2
post==2019.4.13
protobuf==3.11.1
psycopg2==2.8.4
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
python-dateutil==2.6.0
python-dotenv==0.10.3
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==16.7.8
Werkzeug==0.16.0
WTForms==2.2.1
